data_plane/services/local_admin/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyyaml==6.0.1
docker==7.1.0
requests>=2.32.0,<3
websockets==12.0
